requirements.txt

pandas
scikit-learn
matplotlib
streamlit
langchain
langchain-openai
faiss-cpu
tiktoken
uv
python-dotenv
